requirements.txt
bs4==0.0.1
nextcord==2.4.2
urlextract==1.8.0
asyncio==3.4.3
requests==2.28.2
Pillow==9.5.0
python-dotenv==1.0.0
